refactor winit common
